requirements.txt
asgiref==3.8.1
boto3==1.34.123
botocore==1.34.123
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.8.0rc2
dj-database-url==2.2.0
Django==5.0.6
django-cors-headers==4.3.1
django-ses==4.1.0
django-storages==1.14.3
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
gunicorn==22.0.0
idna==3.7
jmespath==1.0.1
oauthlib==3.2.2
packaging==24.1
psycopg==3.1.19
psycopg-binary==3.1.19
# psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.1
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
uuid==1.30
